requirements.txt
flask==3.1.0
flask-cors==5.0.0
google-generativeai
python-dotenv